allow temporary expressions and lvalues in `wite!`
This change enables you to pass bare identifier as
a writer in `wite!` macro as well as temporary exressions
(involving temporary borrows too). So now this
macro acceps everything that `std::write!` does.

This is achieved by an `&mut self -> &mut Self` extension
method (which adds `&mut` only when necessary, using auto-ref)
and a match block, which enables temporaries to live
for the whole macro invocation.