capture `pint!` output on `cfg(test)`. fixes #1
This is actually a workaround, not a proper fix,
since it has some limitations -- when the
macro is called indirectly -- by a different crate than
a tested one, the output won't be captured.

This commit also adds a test that
checks the `pint!` and `perr!` output.

Bumping major version, since it changes
observable behaviour of the library.