Improve performance and add lazy_fomat macro
Instead of calling `write!` for each interpolated argument,
now the whole format-expression is wrapped inside a struct
with a Display::fmt implementation. This makes a fmt::Formater
available to us, so we can call Display/Debug::fmt implementations
of interpolated expressions directly. This results in about
20% speedup in larger format-expressions (because we don't have to
go through format_args machinery all the time). Unfortunately,
this also slows down a little bit short format-expression
(with 0 or 1 interpolated expressions), but this seems
like a reasonable compromise.

This commit also adds a DisplayFn struct and `lazy_format`
macro that uses the same syntax as `fomat` but returns
a displayable struct instead of String.